KD: Some fixes related to writing of the Process module manual and "Sequential reactors" calculation mode
git-svn-id: svn+ssh://savannah01.psi.ch/repos/gems/GEMSELEKTOR/trunk@199 7d9feeff-4400-0410-95ed-9cb8b92a518d

+++ gems2/mods/submod/ms_system.h
@@ -194,7 +194,7 @@ protected:
     void systbc_calc( int mode );
     void PHbcalc( double *MsysC, double *MaqC, double *R1C,
                   double *VaqC, double *VsysC );
-    void stbal( int N, int L, float *Smatr, float *DCstc, double *ICm );
+    void stbal( int N, int L, float *Smatr, double *DCstc, double *ICm );
     // mark data
     void mark_ic_to_bc();
     //void mark_ph_to_dc();
